staging.bbclass: fix for TSS about the existence of lowerdir manifest
The current codes are assuming only one manifest is available. But in TSS,
it's possible that lowerdir has created a new manifest. For example,

A depends on B, and B has some changes, that is B.hash1 -> B.hash2.
In such case, a non-tss tasks updates lowerdir to remove the old B.hash1 and
create B.hash2. The overlayerfs mechanism leads to the existence of both B.hash1
and B.hash2 for a later tss task. And that tss task removes B.hash1 and tries
to create B.hash2, and found that it has already there, thus a 'file/direcotry
exists' error.

To fix this, in case of tss, just remove the B.hash2.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>