time: fix runtime version from UNKNOWN to 1.9
The runtime version of time is determined by:
build-aux/git-version-gen .tarball-version

But there's no .tarball-version in the tarball. And time's last
commit dates back to 20210105, it's very unlikely for them to
generate a new tarball. So we add this file manually with ${PV}
as the content.

Before the fix, time --version:
time (GNU Time) UNKNOWN

After the fix, time --version:
time (GNU Time) 1.9

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>